requirements.txt

langchain
langchain-community
langchain-ollama
langchain-core


langchain-openai
openai

langchain-anthropic


langchain-google-genai
google-generativeai

langchain-groq



langchain-huggingface
transformers
huggingface-hub


python-dotenv


numpy
scikit-learn